Move localizations
